+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Coming Next...
 
 - chore: Add unit test
+- feature: Provider DigitalOcean support choosing region/size/image when
+  creating a droplet
+- feature: Provider DigitalOcean Dynamic detect new created server's ssh port is
+  connectable.
 
 ## 0.2.2